Revert readinsystemdatafromexcel and allow other directories
Reverted mistake adding feature to read_in_system_data_from_excel (the last update corrected an error but introduced a new one by assuming a change that is not part of the default). Also add a feature to ensure FESTIV reads in other directories. This may come in handy for example, if users want to have separate file for internal mods or other code to call on.